backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-genai
google-auth
opencv-python-headless
numpy
Pillow
